+++ src/views/StarfieldData.tsx
@@ -1,0 +1,13 @@
+import * as React from 'react';
+import { ControlLabel, FormGroup, Panel, HelpBlock } from 'react-bootstrap';
+import { useTranslation } from 'react-i18next';
+import { useSelector } from 'react-redux';
+
+export default function StarfieldData() {
+  return (
+    <form>
+      <Panel>
+      </Panel>
+    </form>
+  );
+}